{"schema":"decodex/commit/1","summary":"Split app-server activity modules","authority":"manual"}
