+++ apps/decodex/src/agent/app_server/activity.rs
@@ -2,13 +2,14 @@ mod child;
 mod payload;
 mod protocol;
 
-use std::time::Duration;
-
-pub(crate) use self::protocol::protocol_activity_idle_timeout;
 pub(super) use self::{
 	child::ChildActivityAccumulator, payload::redact_identifier,
-	protocol::ProtocolActivityAccumulator,
+	protocol::accumulator::ProtocolActivityAccumulator,
 };
+
+use std::time::Duration;
+
+use crate::state::ProtocolActivitySummary;
 
 const CHILD_BUCKET_MODEL: &str = "Model";
 const WAITING_REASON_MODEL_EXECUTION: &str = "model_execution";
@@ -36,6 +37,13 @@ const OUTPUT_TOKEN_KEYS: &[&str] = &[
 	"total_output_tokens",
 	"totalOutputTokens",
 ];
+
+pub(crate) fn protocol_activity_idle_timeout(
+	protocol_activity: Option<&ProtocolActivitySummary>,
+	base_timeout: Duration,
+) -> Duration {
+	protocol::protocol_activity_idle_timeout(protocol_activity, base_timeout)
+}
 
 fn duration_seconds_i64(duration: Duration) -> i64 {
 	i64::try_from(duration.as_secs()).unwrap_or(i64::MAX)
